feat(web): update new endpoint
