+++ .changeset/great-shirts-punch.md
@@ -1,0 +1,5 @@
+---
+"@yukinu/web": patch
+---
+
+update new endpoint
