+++ apps/web/app/(main)/account/wishlist/page.tsx
@@ -11,7 +11,7 @@ import { getQueryClient, HydrateClient, trpc } from '@/lib/trpc/rsc'
 export const dynamic = 'force-dynamic'
 
 export default function AccountWishlistPage() {
-  void getQueryClient().prefetchQuery(trpc.user.wishlist.queryOptions({}))
+  void getQueryClient().prefetchQuery(trpc.wishlist.get.queryOptions({}))
 
   return (
     <HydrateClient>
