+++ package.json
@@ -51,12 +51,12 @@
     "prod": {
       "@tanstack/react-query": "^5.90.16",
       "nuqs": "^2.8.6",
-      "zod": "4.3.4"
+      "zod": "4.3.5"
     },
     "dev": {
       "@types/node": "^25.0.3",
       "dotenv-cli": "^11.0.0",
-      "tsdown": "^0.19.0-beta.1",
+      "tsdown": "^0.19.0-beta.2",
       "typescript": "^5.9.3"
     },
     "react": {
