+++ packages/db/drizzle/meta/_journal.json
@@ -5,8 +5,8 @@
     {
       "idx": 0,
       "version": "7",
-      "when": 1767430790432,
-      "tag": "0000_stormy_marauders",
+      "when": 1767528192860,
+      "tag": "0000_grey_longshot",
       "breakpoints": true
     }
   ]
